refactor(frontend): extract PlantingPlans row/selection helpers into plantingPlansUtils
Move the pure helpers from pages/PlantingPlans.tsx — date/area value
coercion, hierarchy selection normalization, bed cell resolution, and
the mobile create-form builders — plus the PlantingPlanRow and
MobileCreateFormState types verbatim into pages/plantingPlansUtils.ts.
The page keeps re-exporting the helpers used by existing tests. No
behavior changes.
