+++ frontend/src/pages/PlantingPlans.tsx
@@ -61,9 +61,8 @@ import { AlertSnackbar } from "../components/feedback/AlertSnackbar";
 import {
   plantingPlanAPI,
   type PlantingPlan,
-  type Bed,
 } from "../api/api";
-import type { CultivationType, Field } from "../api/types";
+import type { CultivationType } from "../api/types";
 import { extractApiErrorMessage } from "../api/errors";
 import {
   formatLocalizedNumber,
@@ -75,7 +74,6 @@ import {
   EditableDataGrid,
   createSingleSelectColumn,
   getCalculatedColumnProps,
-  type EditableRow,
   type DataGridAPI,
   type EditableDataGridCommandApi,
   copyRowsToClipboard,
@@ -112,31 +110,35 @@ export {
 } from "../components/planting-plans/areaHierarchySelection";
 
 export { buildAreaColumnHeaderLabel } from "./plantingPlansUtils";
+export {
+  buildMobileCreateForm,
+  getVisibleMobileRows,
+  normalizeSelectionAfterBedChange,
+  normalizeSelectionAfterFieldChange,
+  normalizeSelectionAfterLocationChange,
+  resolveBedCellValue,
+} from './plantingPlansUtils';
+import {
+  areRowsSemanticallyEqual,
+  buildMobileCreateForm,
+  createEmptyMobileCreateForm,
+  getVisibleMobileRows,
+  isAutoAreaRequest,
+  normalizeSelectionAfterBedChange,
+  resolveBedCellValue,
+  toAreaNumericValue,
+  toDateKey,
+  toOptionalNumber,
+  toOptionalString,
+  type MobileCreateFormState,
+  type PlantingPlanRow,
+} from './plantingPlansUtils';
 
 const DATA_GRID_HEADER_LABEL_SX = { fontWeight: 600 };
 
 /**
  * Row data type for Data Grid
  */
-interface PlantingPlanRow extends PlantingPlan, EditableRow {
-  id: number;
-  isNew?: boolean;
-  location_id?: number;
-  field_id?: number;
-  area_m2?: number;
-  plants_count?: number | null; // UI-only derived field
-  note_attachment_count?: number;
-}
-
-interface MobileCreateFormState {
-  culture: string;
-  bed: string;
-  cultivation_type: CultivationType | "";
-  planting_date: string;
-  area_m2: string;
-  plants_count: string;
-  notes: string;
-}
 
 interface AreaValidationDialogState {
   rowId: number;
@@ -232,237 +234,7 @@ const CultivationTypeEditCell = memo(function CultivationTypeEditCell({
   && previous.placeholder === next.placeholder
 ));
 
-const toDateKey = (value: unknown): number | null => {
-  const buildDateKey = (year: number, month: number, day: number): number | null => {
-    const date = new Date(Date.UTC(year, month - 1, day));
-    if (
-      date.getUTCFullYear() !== year ||
-      date.getUTCMonth() !== month - 1 ||
-      date.getUTCDate() !== day
-    ) {
-      return null;
-    }
-    return date.getTime();
-  };
-  if (value instanceof Date && !Number.isNaN(value.getTime())) {
-    return Date.UTC(value.getFullYear(), value.getMonth(), value.getDate());
-  }
-  if (typeof value !== "string") {
-    return null;
-  }
-  const trimmed = value.trim();
-  const isoMatch = /^(\d{4})-(\d{1,2})-(\d{1,2})/.exec(trimmed);
-  if (isoMatch) {
-    const [, year, month, day] = isoMatch;
-    return buildDateKey(Number(year), Number(month), Number(day));
-  }
-  const germanMatch = /^(\d{1,2})\.(\d{1,2})\.(\d{4})$/.exec(trimmed);
-  if (germanMatch) {
-    const [, day, month, year] = germanMatch;
-    return buildDateKey(Number(year), Number(month), Number(day));
-  }
-  const parsed = new Date(trimmed);
-  if (Number.isNaN(parsed.getTime())) {
-    return null;
-  }
-  return Date.UTC(parsed.getFullYear(), parsed.getMonth(), parsed.getDate());
-};
-
-const toAreaNumericValue = (value: unknown, locale: string): number | null => {
-  const directNumeric = toNumericValue(value);
-  if (directNumeric !== null) {
-    return directNumeric;
-  }
-  if (typeof value !== "string") {
-    return null;
-  }
-  const normalized = value
-    .replace("m²", "")
-    .replace(/\s+/g, " ")
-    .trim();
-  if (normalized.length === 0) {
-    return null;
-  }
-  const localized = parseLocalizedNumber(normalized, locale);
-  return localized === null ? null : localized;
-};
-
-const isAutoAreaRequest = (value: unknown, maxKeyword: string): boolean => {
-  if (value === null || value === undefined) {
-    return true;
-  }
-  if (typeof value !== "string") {
-    return false;
-  }
-  const normalized = value.trim().toLowerCase();
-  return normalized === "" || normalized === maxKeyword.trim().toLowerCase() || normalized === "maximum";
-};
-
-const toOptionalString = (value: unknown): string | undefined =>
-  typeof value === "string" ? value : undefined;
-
-const toOptionalNumber = (value: unknown): number | undefined => {
-  if (typeof value === "number" && Number.isFinite(value)) {
-    return value;
-  }
-  if (typeof value === "string" && /^\d+$/.test(value.trim())) {
-    return Number(value.trim());
-  }
-  return undefined;
-};
-
-interface HierarchySelectionRow {
-  location_id?: number;
-  field_id?: number;
-  bed?: number | string;
-}
-
-export const normalizeSelectionAfterLocationChange = (
-  row: HierarchySelectionRow,
-  nextLocationId: number,
-  fields: Field[],
-  beds: Bed[],
-): HierarchySelectionRow => {
-  const selectedField =
-    typeof row.field_id === "number"
-      ? fields.find((field) => field.id === row.field_id)
-      : undefined;
-  const nextFieldId =
-    selectedField && selectedField.location === nextLocationId
-      ? selectedField.id
-      : undefined;
-
-  const selectedBed =
-    typeof row.bed === "number"
-      ? beds.find((bed) => bed.id === row.bed)
-      : undefined;
-  const selectedBedField =
-    selectedBed && typeof selectedBed.field === "number"
-      ? fields.find((field) => field.id === selectedBed.field)
-      : undefined;
-  const nextBedId =
-    selectedBed &&
-    selectedBedField &&
-    selectedBedField.location === nextLocationId &&
-    (!nextFieldId || selectedBed.field === nextFieldId)
-      ? selectedBed.id
-      : 0;
-
-  return {
-    ...row,
-    location_id: nextLocationId,
-    field_id: nextFieldId,
-    bed: nextBedId,
-  };
-};
-
-export const normalizeSelectionAfterFieldChange = (
-  row: HierarchySelectionRow,
-  nextFieldId: number,
-  fields: Field[],
-  beds: Bed[],
-): HierarchySelectionRow => {
-  const selectedField = fields.find((field) => field.id === nextFieldId);
-  const selectedBed =
-    typeof row.bed === "number"
-      ? beds.find((bed) => bed.id === row.bed)
-      : undefined;
-  const nextBedId =
-    selectedBed && selectedBed.field === nextFieldId ? selectedBed.id : 0;
-
-  return {
-    ...row,
-    location_id: selectedField?.location ?? row.location_id,
-    field_id: nextFieldId,
-    bed: nextBedId,
-  };
-};
-
-export const normalizeSelectionAfterBedChange = (
-  row: HierarchySelectionRow,
-  nextBedId: number,
-  fields: Field[],
-  beds: Bed[],
-): HierarchySelectionRow => {
-  const selectedBed = beds.find((bed) => bed.id === nextBedId);
-  const selectedField = selectedBed
-    ? fields.find((field) => field.id === selectedBed.field)
-    : undefined;
-  return {
-    ...row,
-    location_id: selectedField?.location ?? row.location_id,
-    field_id: selectedBed?.field ?? row.field_id,
-    bed: nextBedId,
-  };
-};
-
 export { buildBedDisplayLabel } from "./plantingPlansUtils";
-
-export const resolveBedCellValue = (
-  value: unknown,
-  row: HierarchySelectionRow,
-): number => {
-  const cellBedId = toNumericValue(value);
-  if (cellBedId !== null) {
-    return cellBedId;
-  }
-
-  const rowBedId = toNumericValue(row.bed);
-  return rowBedId ?? 0;
-};
-
-const createEmptyMobileCreateForm = (): MobileCreateFormState => ({
-  culture: "",
-  bed: "",
-  cultivation_type: "",
-  planting_date: "",
-  area_m2: "",
-  plants_count: "",
-  notes: "",
-});
-
-export const getVisibleMobileRows = (
-  rows: PlantingPlanRow[],
-): PlantingPlanRow[] => rows.filter((row) => !row.isNew);
-
-const areRowsSemanticallyEqual = (
-  previousRows: PlantingPlanRow[],
-  nextRows: PlantingPlanRow[],
-): boolean => {
-  if (previousRows.length !== nextRows.length) {
-    return false;
-  }
-
-  return previousRows.every((previousRow, index) => {
-    const nextRow = nextRows[index];
-    return (
-      previousRow.id === nextRow.id &&
-      previousRow.bed === nextRow.bed &&
-      previousRow.area_m2 === nextRow.area_m2 &&
-      previousRow.plants_count === nextRow.plants_count &&
-      previousRow.notes === nextRow.notes &&
-      previousRow.culture === nextRow.culture &&
-      previousRow.cultivation_type === nextRow.cultivation_type &&
-      previousRow.planting_date === nextRow.planting_date &&
-      previousRow.harvest_date === nextRow.harvest_date &&
-      previousRow.harvest_end_date === nextRow.harvest_end_date
-    );
-  });
-};
-
-export const buildMobileCreateForm = (
-  prefill?: { cultureId?: number | null; bedId?: number | null },
-): MobileCreateFormState => {
-  const baseForm = createEmptyMobileCreateForm();
-
-  return {
-    ...baseForm,
-    culture:
-      typeof prefill?.cultureId === "number" ? String(prefill.cultureId) : "",
-    bed: typeof prefill?.bedId === "number" ? String(prefill.bedId) : "",
-    cultivation_type: "pre_cultivation",
-  };
-};
 
 function PlantingPlans() {
   const { t } = useTranslation(["plantingPlans", "common"]);
